requirements.txt
psycopg2-binary==2.9.7
sqlalchemy==2.0.20